+++ vpmobil/io.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 import xml.etree.ElementTree as XML
 
-from .parser import VpDay
+from .models import VpDay
 
 def parsefromfile(pfad: Path) -> VpDay:
     """